Cleanup of encrypted packet reading and more SSH channel messages
SSHChannel is now stored weakly in the SSH context, instead of holding references to its functions.
While decrypting data, the header is decrypted first and stored. If not enough data is present, it will be reused when more data arrives
Unencrypted packets, for the handshake, are put in a separate and simpler function